+++ src/test/scala/cognite/spark/v1/RawTableRelationTest.scala
@@ -435,57 +435,71 @@ class RawTableRelationTest
       assert(nestedStruct.schema != null)
       nestedStruct.schema.fieldNames.toSeq.loneElement shouldBe "foo"
       nestedStruct.toSeq.loneElement shouldBe 123L
-
-      val arrayOfStruct = struct.getSeq[Row](struct.fieldIndex("array_of_struct"))
-      val structInArray = arrayOfStruct.loneElement
-
-      assert(structInArray.schema != null)
-      structInArray.schema.fieldNames.toSeq.loneElement shouldBe "foo"
-      structInArray.toSeq.loneElement shouldBe 123L
-    } finally {
-      writeClient.rawRows(database, table).deleteById(key)
     }
   }
 
-  it should "create the table with ensureParent option" in {
-    val database = "testdb"
-    val table = "ensureParent-test-" + shortRandomString()
+  it should "be able to duplicate a table with a large number of columns(384)" taggedAs WriteTest in {
+    val source = spark.read
+      .format("cognite.spark.v1")
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("database", "testdb")
+      .option("table", "MegaColumnTable")
+      .option("inferSchema", "true")
+      .option("inferSchemaLimit", "100")
+      .load()
 
-    // remove the DB to be sure
-    try {
-      writeClient.rawTables(database).deleteById(table)
-    } catch {
-      case _: CdpApiException => () // Ignore
-    }
+    val dest = spark.read
+      .format("cognite.spark.v1")
+      .schema(source.schema)
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("database", "testdb")
+      .option("table", "MegaColumnTableDuplicate")
+      .option("inferSchema", "true")
+      .option("inferSchemaLimit", "100")
+      .load()
+    dest.createTempView("megaColumnTableDuplicateTempView")
 
-    val key = shortRandomString()
+    source
+      .select("*")
+      .write
+      .insertInto("megaColumnTableDuplicateTempView")
 
-    try {
-      val source = spark.sql(s"""select
-           |  '$key' as key,
-           |  123 as something
-           |""".stripMargin)
-      val destination = spark.read
-        .format("cognite.spark.v1")
-        .schema(source.schema)
-        .option("apiKey", writeApiKey)
-        .option("type", "raw")
-        .option("database", database)
-        .option("table", table)
-        .option("rawEnsureParent", "true")
-        .load()
-      destination.createTempView("ensureParent_test")
-      source
-        .select(destination.columns.map(c => col(c)): _*)
-        .write
-        .insertInto("ensureParent_test")
+    assert(source.count() == dest.count())
+  }
 
-    } finally {
-      try {
-        writeClient.rawTables(database).deleteById(table)
-      } catch {
-        case _: CdpApiException => () // Ignore
-      }
-    }
+  it should "be treated as a 'select *' when the column names combined, exceeds the character limit of 200" taggedAs WriteTest in {
+    val source = spark.read
+      .format("cognite.spark.v1")
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("database", "testdb")
+      .option("table", "MegaColumnTable")
+      .option("inferSchema", "true")
+      .option("inferSchemaLimit", "100")
+      .load()
+
+    val dest = spark.read
+      .format("cognite.spark.v1")
+      .schema(source.schema)
+      .option("apiKey", writeApiKey)
+      .option("type", "raw")
+      .option("database", "testdb")
+      .option("table", "MegaColumnTableDuplicate2")
+      .option("inferSchema", "true")
+      .option("inferSchemaLimit", "100")
+      .load()
+    dest.createTempView("megaColumnTableDuplicate2TempView")
+
+    val allColumnNamesAsString = source.schema.fields.map(_.name).mkString(",")
+    assert(allColumnNamesAsString.length > 200)
+
+    source
+      .select(source.schema.fields.map(f => col(f.name)): _*)
+      .write
+      .insertInto("megaColumnTableDuplicate2TempView")
+
+    assert(source.count() == dest.count())
   }
 }