+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.4.44
+
+## Fixes
+
+* Fixed a bug where filtering on many RAW columns would create too long URL requests.
+
+# 1.4.43
+
+## Fixes
+
+* Specify timeStamp for upper bound when retrieving dataPoints.
+
 # 1.4.42
 
 ## Enhancements